requirements-dev.txt
# Development requirements for the Meld & RAG system
# Additional packages for development and debugging

# Development tools
ipython>=8.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
pylint>=2.12.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.6.0
pytest-flask>=1.2.0

# Debugging
pdb++>=0.10.0
ipdb>=0.13.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Profiling and monitoring
memory-profiler>=0.60.0
line-profiler>=3.5.0
py-spy>=0.3.0

# Database tools
sqlite-utils>=3.25.0

# API development
httpie>=3.0.0
requests-toolbelt>=0.9.0

# OpenAI Integration
openai>=1.0.0

# Environment management
python-dotenv>=0.19.0
